### 2019-12-12 #1953 @whedon
[PRE REVIEW]: DAFSA: a Python library for Deterministic Acyclic FiniteState Automata
**Submitting author:** @tresoldi (<a href="http://orcid.org/0000-0002-2863-1467">Tiago Tresoldi</a>)
**Repository:** <a href="https://github.com/tresoldi/dafsa" target ="_blank">https://github.com/tresoldi/dafsa</a>
**Version:** v0.5.1
