### 2019-12-12 #1953 @kyleniemeyer
Hi @tresoldi, could you review the references in your article and add DOIs for those missing?
