### 2019-12-12 180c59d
Adding DOIs
