+++ manuscript/paper.bib
@@ -23,7 +23,8 @@
   volume={40},
   pages={31--55},
   year={1985},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/0304-3975(85)90157-4}
 }
 
 @article{Ciura:2001,
@@ -34,7 +35,8 @@
   number={11},
   pages={1077--1090},
   year={2001},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/spe.402}
 }
 
 @article{Daciuk:2000,
@@ -56,7 +58,8 @@
   number={1},
   pages={15--30},
   year={1993},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://doi.org/10.1002/spe.4380230103}
 }
 
 @electronic{Havon:2011,
