+++ manuscript/paper.bib
@@ -59,7 +59,7 @@
   pages={15--30},
   year={1993},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1002/spe.4380230103}
+  doi={10.1002/spe.4380230103}
 }
 
 @electronic{Havon:2011,
