### 2019-12-12 #1953 @tresoldi
Thank you, everyone!

I've pushed a `paper.bib` with the DOIs (commit https://github.com/tresoldi/dafsa/commit/1e3b555f00c90809d172fa45a5bdf34adf680163) and the preview is working ([here](https://res.cloudinary.com/hju22ue2k/image/upload/v1576182013/hdshcpv6jw19uz8so5fl.pdf)).
