### 2019-12-28 #1953 @geoffbacon
> @desilinguist, @jkahn, @ionlights, @mbod, @geoffbacon, are any 2 of you willing to contribute a review for this JOSS submission?

@gkthiruvathukal I'd be happy to review!
