### 2019-12-28 #1986 @whedon
[REVIEW]: DAFSA: a Python library for Deterministic Acyclic FiniteState Automata
**Submitting author:** @tresoldi (<a href="http://orcid.org/0000-0002-2863-1467">Tiago Tresoldi</a>)
**Repository:** <a href="https://github.com/tresoldi/dafsa" target ="_blank">https://github.com/tresoldi/dafsa</a>
**Version:** v1.0.zenodo
