### 2020-01-22 tresoldi/dafsa#3 @desilinguist
Suggestions for improving documentation
The documentation is really nice! I have a few suggestion to make it even better:

1. Provide a clear statement-of-need like you do in the JOSS paper.
2. Provide a full walkthrough of the first example you provide in the "How to use" section. For example, it is not clear what all the numbers and symbols mean. It would be nice if you take one line of the example output and document what each symbol means.
3. It would also be nice if you could provide a fully worked-through realistic example. For example, the phonemes example is pretty nice but it isn't really clear (other than to phoneticians) what the various strings mean.
