### 2020-01-22 tresoldi/dafsa#4 @desilinguist
Please provide explicit list of dependencies
1. It would be great if you could explicitly list the dependencies in the README.
2. Please use a `requirements.txt` file rather than just listing the dependencies in `setup.py`.
