### 2020-01-22 tresoldi/dafsa#5 @desilinguist
Please consider creating a conda package 
It would be nice if you could create a [conda](https://conda.io/en/latest:) package and release it on anaconda.org. That would make it very easy for folks like me who prefer to use `conda` over `pip`.
