### 2020-01-22 #1986 @desilinguist
I think the package and the paper both look great! I have some suggestions to improve the [documentation](https://github.com/tresoldi/dafsa/issues/3), the [dependencies](https://github.com/tresoldi/dafsa/issues/4), and the [packaging](https://github.com/tresoldi/dafsa/issues/5).
