### 2020-01-22 tresoldi/dafsa#4 @geoffbacon
I'm also in favour of these two suggestions. Indeed, the fact that dafsa only has one third party dependency (networkx) is a strong positive. Potential users will have more confidence that the installation will "just work".
