### 2020-01-22 #1986 @geoffbacon
Thanks for the great package! I can definitely see myself using this in research and classes. I've left comments and new issues with suggestions.
