+++ docs/readme.rst
@@ -29,12 +29,16 @@ estimate the sampling probability of unobserved sequences.
 
 This data structure is a special case of a finite state recognizer that
 acts as a deterministic finite state automaton, as it recognizes all and
-only the sequences it was built upon. Frequently used in computer
+tonly the sequences it was built upon. It is frequently used in computer
 science for the space-efficient storage of sets of sequences without
 common compression techniques, such as dictionary or entropy types, or
-without probabilistic data structures, such as Bloom filters, the
-automata generated by this library are intended for linguistic
-exploration, and extend published models by allowing to approximate
+without probabilistic data structures, such as Bloom filters.
+
+The
+automata generated by this library are intended for
+usage by linguists in studies on morphology and formal grammars, allowing
+for faster, easier, and simpler generation of visualizations. The
+library also extends published models by allowing to approximate
 probability of random observation by carrying information on the weight
 of each graph edge.
 
