+++ docs/quickstart.rst
@@ -10,14 +10,15 @@ A minimal usage is shown here:
    >>> from dafsa import DAFSA
    >>> d = DAFSA(["tap", "taps", "top", "tops", "dibs"])
    >>> print(d)
-   DAFSA with 7 nodes and 8 edges (5 sequences)
-     +-- #0: 0(#1:<d>/1|#5:<t>/4) [('t', 5), ('d', 1)]
-     +-- #1: n(#2:<i>/1) [('i', 2)]
-     +-- #2: n(#3:<b>/1) [('b', 3)]
-     +-- #3: n(#4:<s>/3) [('s', 4)]
+   DAFSA with 8 nodes and 9 edges (4 inserted sequences)
+     +-- #0: 0(#1/4:<d>/1|#5/4:<t>/3) [('d', 1), ('t', 5)]
+     +-- #1: n(#2/1:<i>/1) [('i', 2)]
+     +-- #2: n(#3/1:<b>/1) [('b', 3)]
+     +-- #3: F(#4/3:<s>/2) [('s', 4)]
      +-- #4: F() []
-     +-- #5: n(#6:<a>/2|#6:<o>/2) [('o', 6), ('a', 6)]
-     +-- #6: n(#3:<p>/4) [('p', 3)]
+     +-- #5: n(#6/3:<a>/2|#9/3:<o>/1) [('a', 6), ('o', 9)]
+     +-- #6: n(#3/2:<p>/2) [('p', 3)]
+     +-- #9: n(#4/1:<p>/1) [('p', 4)]
 
 Note how the resulting graph includes the 5 training sequences, with one
 starting node (#0) that advances with either a ``t`` (observed four
@@ -99,9 +100,133 @@ specified in source by using spaces as token delimiters:
 .. figure:: https://raw.githubusercontent.com/tresoldi/dafsa/master/figures/phonemes.png
    :alt: Phoneme example
 
-DAFSA structure can be exported as:
+DAFSA structures can be exported in PDF, SVG, GLM, and DOT formats.
 
-- PDF
-- SVG
-- GLM
-- DOT
+Walkthrough example
+-------------------
+
+The most basic DAFSA is one expressing a single string, which links
+the sequence of characters. As such, let's start by creating a DAFSA for
+the string ``tap``, exploring the results.
+
+.. code:: python
+
+   >>> from dafsa import DAFSA
+   >>> d1 = DAFSA(["tap"])
+   >>> print(d1)
+   DAFSA with 4 nodes and 3 edges (1 inserted sequences)
+     +-- #0: 0(#1/1:<t>/1) [('t', 1)]
+     +-- #1: n(#2/1:<a>/1) [('a', 2)]
+     +-- #2: n(#3/1:<p>/1) [('p', 3)]
+     +-- #3: F() []
+
+When printing the structure, we are informed that it contains four nodes:
+the three chracters ``t``, ``a``, and ``p`` from our string, plus a shared
+final node. In order, skipping over information we will see later, we have:
+
+The textual presentation given by ``print()`` lists all nodes, showing, in
+order, the node index (in format ``#i:``, where `i` is the index), an
+unambiguous textual representation (such as `#1/1:<t>/1`) and a list of
+edges involved, conceived as a more human-friendly way of rendering the
+information in the textual representation. In this example, we find out
+that:
+
+- A node of index 0 (``#0``) and type ``0`` (that is, a start symbol),
+  which, as expressed by the list ``[('t', 1)]`` can only transition
+  ("advance") with the character `t` moving to the node of index 1.
+- A node of index 1 and type ``n`` (that is, a "normal" node which is neither
+  a start or end symbol), which can only transition with the character
+  `a` towards the node of index 2.
[…]
