+++ setup.py
@@ -7,6 +7,10 @@ LOCAL_PATH = pathlib.Path(__file__).parent
 # The text of the README file
 README_FILE = (LOCAL_PATH / "README.md").read_text()
 
+# Load requirements, so they are listed in a single place
+with open('requirements.txt') as fp:
+    install_requires = [dep.strip() for dep in fp.readlines()]
+
 # This call to setup() does all the work
 setup(
     name="dafsa",
@@ -36,7 +40,7 @@ setup(
         "directed acyclic word graph",
     ],
     include_package_data=True,
-    install_requires=["networkx"],
+    install_requires=install_requires,
     entry_points={"console_scripts": ["dafsa=dafsa.__main__:main"]},
     test_suite="tests",
     tests_require=[],
