+++ MANIFEST.in
@@ -1,4 +1,5 @@
 include README.md
 include AUTHORS.md
 include LICENSE
+include requirements.txt
 include resources/*
