### 2020-02-01 3ca7a08
Add contributing guidelines.
