+++ MANIFEST.in
@@ -1,5 +1,6 @@
 include README.md
 include AUTHORS.md
+include CONTRIBUTING.md
 include LICENSE
 include requirements.txt
 include resources/*
