### 2020-02-01 tresoldi/dafsa#5 @tresoldi
Sure, and it would be a good practice, I've never created a conda package before.

I'll get to it once I'm finished with the other issues, and comment again here.
