### 2020-02-01 tresoldi/dafsa#3 @tresoldi
First item is done, a draft of the second is on-going ([here](https://github.com/tresoldi/dafsa/blob/joss/docs/quickstart.rst)), and the third one is pending.

Unfortunately I found a bug while writing the documentation, of the kind that seems to be only triggered in specific situations and thus a bit harder to debug. I will work on it tomorrow (one more example that documentation is good for finding bugs...).
