### 2020-02-01 05b58cb
Fix bug where node finality is not considered in equality comparison.
