### 2020-02-02 2b3d6df
Add extended documentation and regenerate figures
