+++ docs/quickstart.rst
@@ -10,20 +10,22 @@ A minimal usage is shown here:
    >>> from dafsa import DAFSA
    >>> d = DAFSA(["tap", "taps", "top", "tops", "dibs"])
    >>> print(d)
-   DAFSA with 8 nodes and 9 edges (4 inserted sequences)
-     +-- #0: 0(#1/4:<d>/1|#5/4:<t>/3) [('d', 1), ('t', 5)]
+   DAFSA with 8 nodes and 9 edges (5 inserted sequences)
+     +-- #0: 0(#1/5:<d>/1|#5/5:<t>/4) [('d', 1), ('t', 5)]
      +-- #1: n(#2/1:<i>/1) [('i', 2)]
      +-- #2: n(#3/1:<b>/1) [('b', 3)]
-     +-- #3: F(#4/3:<s>/2) [('s', 4)]
+     +-- #3: n(#4/1:<s>/1) [('s', 4)]
      +-- #4: F() []
-     +-- #5: n(#6/3:<a>/2|#9/3:<o>/1) [('a', 6), ('o', 9)]
-     +-- #6: n(#3/2:<p>/2) [('p', 3)]
-     +-- #9: n(#4/1:<p>/1) [('p', 4)]
+     +-- #5: n(#6/4:<a>/2|#6/4:<o>/2) [('o', 6), ('a', 6)]
+     +-- #6: n(#7/4:<p>/4) [('p', 7)]
+     +-- #7: F(#4/4:<s>/2) [('s', 4)]
 
 Note how the resulting graph includes the 5 training sequences, with one
-starting node (#0) that advances with either a ``t`` (observed four
+start node (#0) that advances with either a ``t`` (observed four
 times) or a ``d`` symbol (observed a single time), a subsequent node to
-``t`` that only advances with ``a`` and ``o`` symbols (#1), and so on.
+``t`` that only advances with ``a`` and ``o`` symbols (#1), and so on. More
+information on how to interpret the structure of a DAFSA are given in the
+following subsections.
 
 The structure is much clearer with a graphical representation:
 
@@ -42,9 +44,9 @@ cumulative path weight is also returned.
 .. code:: python
 
    >>> d.lookup("tap")
-   (F(#4/5:<s>/3), 10)
+   (F(#4/4:<s>/2), 10)
    >>> d.lookup("tops")
-   (F(), 13)
+   (F(), 12)
    >>> d.lookup("tapap") is None
    True
    >>> d.lookup("ta") is None
@@ -105,9 +107,10 @@ DAFSA structures can be exported in PDF, SVG, GLM, and DOT formats.
 Walkthrough example
 -------------------
 
-The most basic DAFSA is one expressing a single string, which links
-the sequence of characters. As such, let's start by creating a DAFSA for
-the string ``tap``, exploring the results.
+The simplest DAFSAs are capable of expressing only one string, linking its
+sequence of characters from the first to the last. The creation of such a DAFSA
+follows the procedures described above, here demonstrated with the string
+``tap``:
 
 .. code:: python
 
@@ -120,30 +123,32 @@ the string ``tap``, exploring the results.
      +-- #2: n(#3/1:<p>/1) [('p', 3)]
      +-- #3: F() []
 
-When printing the structure, we are informed that it contains four nodes:
-the three chracters ``t``, ``a``, and ``p`` from our string, plus a shared
-final node. In order, skipping over information we will see later, we have:
+The printed textual representation confirms that only one sequence was
+inserted, with three edges (or "transitions", corresponding to the
+three characters in our string) among four nodes (that is, a start node,
+two transitional nodes, and a final node).
+A unique but meaningless index (in this basic example, the numbers
+from zero to three) identifies each node, an explicit textual
+representation (such as
[…]
