### 2020-02-02 83e9a40
Add RST version of the README for PyPI/twine compatibility
