### 2020-02-02 tresoldi/dafsa#4 @tresoldi
@desilinguist and @geoffbacon : implemented and published in the new version 0.6. Please close the issue if it is ok.
