### 2020-02-02 d39d83a
Add resources for the conda package
