### 2020-02-02 tresoldi/dafsa#5 @tresoldi
@desilinguist The package is online: https://anaconda.org/tresoldi/dafsa

Resources for the package were added with commit d39d83a9881ef7c083e593b384fd2772d5f434b5

Let me know if it is ok (as I commented above, it is the first time I build a conda package).
