### 2020-02-03 tresoldi/dafsa#5 @desilinguist
Thank you so much for taking up my suggestion! You might also want to add another badge to the README for the conda package.
