### 2020-02-03 #1986 @desilinguist
All of my suggestions were very promptly addressed by @tresoldi! IMO, the paper can now be accepted.
