### 2020-02-13 #1986 @geoffbacon
@gkthiruvathukal Sorry - I thought having completed the checklist was sufficient to indicate that all the feedback/suggestions had been addressed. The paper/library is ready for acceptance from my point of view.
