### 2020-02-16 #1986 @gkthiruvathukal
@tresoldi Ok, we're on our way toward acceptance. Can you please check that all of the following are done?  I know that you may have done some of these already, but this makes my life easier once I signal to the EICs that we're ready to accept.

Thank you!
