### 2020-02-16 47c0797
Bump up version to 1.0
