+++ README.rst
@@ -195,10 +195,6 @@ Version 0.1:
 Roadmap
 -------
 
-Version 1.0:
-
--  Publish in journal
-
 After 1.0:
 
 -  Preliminary generation of minimal regular expressions matching the
@@ -231,20 +227,20 @@ programme (grant agreement No. \ `ERC Grant
 
 If you use ``dafsa``, please cite it as:
 
-   Tresoldi, Tiago (2019). DAFSA, a a library for computing
-   Deterministic Acyclic Finite State Automata. Version 0.6. Jena.
+   Tresoldi, Tiago (2020). DAFSA, a a library for computing
+   Deterministic Acyclic Finite State Automata. Version 1.0. Jena.
    Available at: https://github.com/tresoldi/dafsa
 
 In BibTeX:
 
 .. code:: bibtex
 
-   @misc{Tresoldi2019dafsa,
+   @misc{Tresoldi2020dafsa,
      author = {Tresoldi, Tiago},
-     title = {DAFSA, a a library for computing Deterministic Acyclic Finite State Automata. Version 0.6.},
+     title = {DAFSA, a a library for computing Deterministic Acyclic Finite State Automata. Version 1.0.},
      howpublished = {\url{https://github.com/tresoldi/dafsa}},
      address = {Jena},
-     year = {2019},
+     year = {2020},
    }
 
 .. |PyPI| image:: https://img.shields.io/pypi/v/dafsa.svg
