### 2020-02-16 fac26ee
Bumping year in docs configuration
