### 2020-02-16 e17ae76
Updating date in JOSS manuscript and metadata
