+++ docs/readme.rst
@@ -97,7 +97,7 @@ If you use ``dafsa``, please cite it as:
 
    Tresoldi, Tiago (2020). DAFSA, a a library for computing
    Deterministic Acyclic Finite State Automata. Version 1.0. Jena.
-   Available at: https://github.com/tresoldi/dafsa
+   DOI: 10.5281/zenodo.3668870
 
 In BibTeX:
 
@@ -108,6 +108,7 @@ In BibTeX:
      title = {DAFSA, a a library for computing Deterministic Acyclic Finite State Automata. Version 1.0},
      howpublished = {\url{https://github.com/tresoldi/dafsa}},
      address = {Jena},
+     doi = {10.5281/zenodo.3668870},
      year = {2020},
    }
 
