+++ README.rst
@@ -229,7 +229,7 @@ If you use ``dafsa``, please cite it as:
 
    Tresoldi, Tiago (2020). DAFSA, a a library for computing
    Deterministic Acyclic Finite State Automata. Version 1.0. Jena.
-   Available at: https://github.com/tresoldi/dafsa
+   DOI: 10.5281/zenodo.3668870
 
 In BibTeX:
 
@@ -240,6 +240,7 @@ In BibTeX:
      title = {DAFSA, a a library for computing Deterministic Acyclic Finite State Automata. Version 1.0.},
      howpublished = {\url{https://github.com/tresoldi/dafsa}},
      address = {Jena},
+     doi = {10.5281/zenodo.3668870},
      year = {2020},
    }
 
