### 2020-02-16 #1986 @tresoldi
Thank you, @gkthiruvathukal ! So that I am not the one checking the boxes:

- The tagged release is `v1.0.zenodo`: https://github.com/tresoldi/dafsa/releases/tag/v1.0.zenodo (the matching package is [on PyPi](https://pypi.org/project/dafsa/) as well, and I'll rebuild the Conda package after publication)
- The release is archived on Zenodo at: https://zenodo.org/record/3668870
- The Zenodo metadata is correct and matching the one of the paper/package metadata (such license, funding, etc.); my name is correct and it is linked to ORCID
- The Zenodo DOI is 10.5281/zenodo.3668870 ([link](https://doi.org/10.5281/zenodo.3668869)) (which was also added to the documentation in the repository and as a badge in the README file)
