### 2020-02-16 3f174af
Add JOSS badge to documentation
